requirements.txt
Flask==2.3.3
PyMySQL==1.1.0
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
Werkzeug==2.3.7
